Adds preclustering by mash distance to the `rake parsnp` task
Can use the new MASH_CUTOFF and MAX_CLUSTER_SIZE parameters to tune the
clustering process. MASH_CUTOFF is the maximum diameter of a cluster in
mash distance units; MAX_CLUSTER_SIZE hard limits the size of a cluster
and can be used to prevent parsnp from trying to align too many genomes.

All of this is intended to boost the core genome size that remains for
parsnp comparisons and SNV distances.

TODO: still need to figure out how to extract the core genome size from
the .ggr files, and add it to the .parsnp.heatmap.json files so it can
show up in the heatmap viz.