Upgrades to `rake parsnp` to handle 1-genome mash clusters
As well as fixes to the dependency graph to avoid unnecessary rebuilds.
Directories have been pulled out of the dependency graph because their
mtime changes whenever a file is added/removed/renamed, which can
unnecessarily rebuild everything within them that lists the directory as
a dependency.

Note that there is now a `rake encounters` task for generating
encounters.tsv files.